+++ src/frontend/web_application/src/modules/discussion/index.js
@@ -1,0 +1,4 @@
+export * from './actions/requestDiscussionIdForParticipants';
+export * from './selectors/discussionDraftSelector';
+export * from './selectors/discussionIdSelector';
+export * from './selectors/discussionSelector';