+++ src/frontend/web_application/src/modules/discussion/selectors/discussionIdSelector.js
@@ -1,0 +1,7 @@
+export const discussionIdSelector = (state, { discussionId }) => {
+  if (!discussionId) {
+    throw new Error('discussionId must be passed when calling discussionIdSelector, e.g discussionIdSelector(state, { discussionId })');
+  }
+
+  return discussionId;
+};