+++ src/frontend/web_application/src/scenes/NewDraft/components/DraftDiscussion/withDraftDiscussion.jsx
@@ -1,0 +1,171 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import sha1 from 'uuid/lib/sha1-browser';
+import bytesToUuid from 'uuid/lib/bytesToUuid';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
+import { createSelector } from 'reselect';
+import { getModuleStateSelector } from '../../../../store/selectors/getModuleStateSelector';
+import { createMessageCollectionStateSelector } from '../../../../store/selectors/message';
+import { draftSelector } from '../../../../modules/draftMessage';
+import { requestDiscussionIdForParticipants } from '../../../../modules/discussion';
+import { requestDiscussion } from '../../../../modules/message';
+import { withIdentities, identityToParticipant, identitiesSelector } from '../../../../modules/identity';
+import { withUser, userSelector } from '../../../../modules/user';
+
+const getParticipants = ({ draftMessage, identities, user }) => {
+  const { user_identities: [identityId], participants = [] } = draftMessage;
+
+  if (!identityId) {
+    return participants;
+  }
+
+  const identity = identities.find(ident => ident.identity_id === identityId);
+  const from = identityToParticipant({ identity, user });
+
+  return [
+    ...participants
+      .filter(part => part.address !== from.address || part.protocol !== from.protocol),
+    from,
+  ];
+};
+
+const getParticipantsHash = ({ participants }) => {
+  if (participants.length === 0) {
+    return undefined;
+  }
+
+  return bytesToUuid(sha1(participants
+    .map(participant => `${participant.address}_${participant.protocol}`)
+    .sort()
+    .join('+')));
+};
+const discussionStateSelector = getModuleStateSelector('discussion');
+const messageStateSelector = getModuleStateSelector('message');
+const draftMessageSelector = (state, { messageId }) =>
+  draftSelector(state, { internalId: messageId });
+const discussionIdSelector = createSelector(
+  [discussionStateSelector, draftMessageSelector, identitiesSelector, userSelector],
+  (discussionState, draftMessage, identities, user) => {
+    if (!draftMessage || !draftMessage.participants || draftMessage.participants.length === 0) {
+      return undefined;
+    }
+
+    const participants = getParticipants({ draftMessage, identities, user });
+
+    const { discussionId } = discussionState.discussionByParticipantsHash[getParticipantsHash({
+      participants,
+    })] || {};
+
+    return discussionId;
+  }
+);
+const discussionSelector = createSelector(
+  [discussionIdSelector, discussionStateSelector],
+  (discussionId, discussionState) => discussionState.discussionsById[discussionId]
+);
+const messageCollectionStateSelector = createMessageCollectionStateSelector(() => 'discussion', discussionIdSelector);
+
+const mapStateToProps = createSelector(
+  [discussionSelector, messageCollectionStateSelector, messageStateSelector, draftMessageSelector],
+  (discussion, messageCollectionState, messageState, draftMessage) => ({
+    draftMessage,
+    discussion,
+    messages: messageCollectionState.messages,
+  })
+);
+
+const requestDraftDiscussion = ({ participants, internalHash }) => async (dispatch) => {
+  const discussionId = await dispatch(requestDiscussionIdForParticipants({
+    participants, internalHash,
+  }));
+  if (discussionId && discussionId.length > 0) {
+    dispatch(requestDiscussion({ discussionId }));
+  }
+};
+
+const mapDispatchToProps = dispatch => bindActionCreators({
+  requestDraftDiscussion,
+}, dispatch);
+
+const connecting = connect(mapStateToProps, mapDispatchToProps);
+
+export const withDraftDiscussion = () => (C) => {
+  @withUser()
+  @withIdentities({ namespace: 'idents' })
+  @connecting
+  class WithDraftDiscussion extends Component {
+    static propTypes = {
+      requestDraftDiscussion: PropTypes.func.isRequired,
+      draftMessage: PropTypes.shape({}),
+      discussion: PropTypes.shape({}),
+      messages: PropTypes.arrayOf(PropTypes.shape({})),
+      // user: PropTypes.shape({}).isRequired,
+      idents: PropTypes.shape({
+        identities: PropTypes.arrayOf(PropTypes.shape({})),
+      }).isRequired,
+      userState: PropTypes.shape({
+        user: PropTypes.shape({}).isRequired,
+      }).isRequired,
+    };
+
+    static defaultProps = {
+      draftMessage: undefined,
+      discussion: undefined,
+      messages: [],
+    }
+
+    componentDidMount() {
+      const { draftMessage } = this.props;
+
+      if (draftMessage) {
+        this.fetchDiscussion();
+      }
+    }
+
+    componentDidUpdate(prevProps) {
+      if (
+        this.props.draftMessage && (
+          !prevProps.draftMessage ||
+          prevProps.draftMessage.participants !== this.props.draftMessage.participants
+        )
+      ) {
+        this.fetchDiscussion();
+      }
+    }
+
+    fetchDiscussion = () => {
+      const {
+        draftMessage,
+        idents: { identities }, userState: { user },
+      } = this.props;
+
+      const participants = getParticipants({ draftMessage, identities, user });
+
+      if (participants.length > 0) {
+        this.props.requestDraftDiscussion({
+          participants,
+          internalHash: getParticipantsHash({ participants }),
+        });
+      }
+    }
+
+    render() {
+      const {
+        requestDraftDiscussion: unused, draftMessage, discussion, messages, ...props
+      } = this.props;
+
+      const draftDiscussionProps = {
+        // draftMessage,
+        discussion,
+        messages,
+      };
+
+      return (
+        <C draftDiscussion={draftDiscussionProps} {...props} />
+      );
+    }
+  }
+
+  return WithDraftDiscussion;
+};